Update README.md
MySQL doesn't work below 5.6.5 due to the migrations having multiple timestamp columns. I found this out while trying to install on Ubuntu 14.04, which ships with 5.5. The software also completely fails to install on the nodejs version (0.10.something) that ships with the base system. Added these warnings to the readme to save others time.