API Cleanup, Refactor articleRouter and many more!
_Note: This was added without linter verification_

This was something I worked on a couple of weeks ago, but forgot to add in.

This refactors the DB helpers into mixins, so that they can be more extendable in the future. Before this we used a decorator fn that would wrap a model, which makes us lose out on a lot of fun things Objection.js gave us, such as

* Changing the default `QueryBuilder` object to query only `active` items
* Adding properties to omit when convering an DB Model to JSON (like the password on the user)
* Using namedFilters and eager queries together to get nested relations
* Reducing the number of queries to the DB
* Make soft delete (`is_active: true/false`) a feature of the `QueryBuilder`

* I've tried leveraging a lot of those fun features now, in the `modelHelpers` file. This has an added side effect of us not needing to test models. So all the Model tests were removed.
* I refactored how the tests were setup and added some more coverage for some edge cases. (Still a long way to go, but good enough for now)
* The ArticleRouter also got a decent refactor.
  * Validations for `POST` and `PUT`
  * Sorting by `name`, `updated_at`, `topic_id` fields
  * Filters via query string such as `?search=` and `?topic_id=` are now possible

To read about the changes in the articleRouter, I would urge reading the articleRouter tests. Granted, that file has gotten humungous, but if you just collapse all the code and read the test description, its sure to give the reader an overall idea as to what the API does.