[Abandoned] UI component rehaul with Storybook, Theme-UI
This was an abandoned idea of building a separate design system for matterwiki with Theme-UI and storybook. I believe that this was going in the right direction,
but may have been overkill for what I want to do with this project. Maybe in the future.

For now, I am leaving this in this commit. It will be replaced with Chakra-UI or something similar!